backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy
psycopg[binary]
python-multipart
openpyxl
motor
python-dotenv
httpx
google-auth
requests
bcrypt
twilio
sendgrid
solapi
